manager/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-manager"
version = "0.1.0"
description = "Kitty 终端任务管理器 - 管理多个 Claude Code 和终端进程"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "User" }
]
keywords = ["kitty", "terminal", "task-manager", "claude", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "textual>=0.40.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
claude-manager = "claude_manager.cli:main"

[project.urls]
Homepage = "https://github.com/user/claude-manager"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
